refactor: reorganize auth middleware into auth domain folder
- Move packages/api/src/middleware/auth.py → packages/api/src/auth/middleware.py
- Update all import paths to use auth.middleware instead of middleware.auth
- Better semantic organization following domain-driven design
- Provides room for future auth modules like permissions.py, roles.py, etc.

This structure makes it more intuitive for developers to find and organize
authentication-related code.
